Промежуточный итог - все перенес в один сервис. все работает, осталось избавиться от лишнего

+++ src/main/java/org/javaacademy/carad/dto/AdvertisementDto.java
@@ -4,6 +4,7 @@ import lombok.NonNull;
 import lombok.Value;
 import org.javaacademy.carad.entity.Car;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Value
@@ -11,7 +12,7 @@ public class AdvertisementDto {
     int id;
     String brandName;
     String color;
-    double price;
+    BigDecimal price;
     String model;
     LocalDate localDate;
 }
